# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Agent Platform API",
    "description": "マルチユーザー対応AIエージェントプラットフォーム",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/execute/{workflow_id}": {
      "post": {
        "tags": [
          "api-trigger"
        ],
        "summary": "Api Execute Workflow",
        "description": "APIキー認証でワークフロー実行.\n\n- X-API-Key ヘッダーでAPIキーを渡す\n- Rate Limiting が適用される\n- スコープ \"workflows:execute\" または \"*\" が必要",
        "operationId": "api_execute_workflow_v1_execute__workflow_id__post",
        "parameters": [
          {
            "name": "workflow_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Workflow Id"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/ApiTriggerRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Request"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiTriggerResponse"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiTriggerErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiTriggerErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiTriggerErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiTriggerErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/rate-limit": {
      "get": {
        "tags": [
          "api-trigger"
        ],
        "summary": "Get Rate Limit Status",
        "description": "現在のRate Limit状態を取得.",
        "operationId": "get_rate_limit_status_v1_rate_limit_get",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Rate Limit Status V1 Rate Limit Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiTriggerErrorResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ApiTriggerErrorResponse",
        "description": "API Trigger エラーレスポンス."
      },
      "ApiTriggerRequest": {
        "properties": {
          "trigger_data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trigger Data"
          }
        },
        "type": "object",
        "title": "ApiTriggerRequest",
        "description": "API Trigger リクエスト."
      },
      "ApiTriggerResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "task_id": {
            "type": "string",
            "title": "Task Id"
          },
          "workflow_id": {
            "type": "string",
            "title": "Workflow Id"
          },
          "rate_limit_remaining": {
            "type": "integer",
            "title": "Rate Limit Remaining"
          }
        },
        "type": "object",
        "required": [
          "status",
          "task_id",
          "workflow_id",
          "rate_limit_remaining"
        ],
        "title": "ApiTriggerResponse",
        "description": "API Trigger レスポンス."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}